configure.ac: clean up --with-qd=foo behavior.
If the user specifies --with-qd=/path/to/somewhere and if we can't
find the headers there, then the build should fail at that point.
This commit implements that change, checking for the headers
immediately and bailing out if they can't be found. The variables like
LDFLAGS and LIBQD_LIBS now only get set in the "success" branch of the
header search.

The ulterior motive for this change is that if PKG_CHECK_MODULES is
introduced, we don't want to be setting the value of LIBQD_LIBS in
two different places.

Issue: https://github.com/fplll/fplll/issues/405