configure.ac: add support for ternary --with-qd=yes/no/path.
When configuring fplll, there is no way to tell the build system that
you want to use the qd library, but that you also want the headers,
libraries, flags, et cetera to be detected by searching the standard
paths. This commit adds support for --with-qd=yes, which is what you
get with the bare --with-qd, by adding a PKG_CHECK_MODULES call and a
manual header- and library-search fallback.

Closes: https://github.com/fplll/fplll/issues/405