configure.ac: act as if --without-qd was passed by default.
The existing cases for the value of the --with-qd flag were written in
such a way that when the user does nothing ($with_qd = ""), it acts as
if you had passed --with-qd="". This does happen to accidentally
enable qd on my system, but it's not the right thing to do; it only
works by chance, and it makes the build a lot less predictable.

Maybe more to the point, users just don't expect fplll to be linked
against libqd when --with-qd wasn't passed. Even more to the point,
this now causes the configure script to bail out when the user doesn't
have qd installed because the script can't find the headers.

This commit sets the default to --with-qd=no explicitly.