configure.ac: remove redundant checks for mp[if]r and gmp.
The configure script currently checks for any version of MPIR, and
then it checks for a version of MPIR newer than v1.0.0. The first
"any version" check is therefore redundant, and pretty much harmless,
but one side-effect of the call to AC_CHECK_LIB is that it appends
"-lfoo" to the LIBS variable when it succeeds. So, we wind up with
LIBS="... -lmpir -lmpir" whenever MPIR >= 1.0 is installed. The same
is true for MPFR and GMP.

In preparation for adding the LIBS variable to fplll's pkg-config
file, this commit removes the redundant check to eliminate the
redundant flags in LIBS.