configure.ac: phase out the custom variable LIBGMP.
The LIBGMP variable in configure.ac (that was propagated throughout
the build system) was intended to hold the linker flags for
multi-precision arithmetic, typically either "-lgmp" or
"-lmpir". However, the two checks for GMP and MPIR are mutually
exclusive, and they both use the AC_CHECK_LIB macro that appends
"-l<library>" to a variable named LIBS when <library> is
found. Therefore LIBS already serves the purpose that LIBGMP serves.

The way this differs from (say) LIBQD_LIBS is that, when pkg-config
is used, the LIBS variable does _not_ have "-lqd" appended to it. So
when pkg-config is used (and AC_SEARCH_LIBS is not), the additional
variable is needed to hold the "-lqd" flag.

Having replaced LIBGMP by LIBS, we are also able to remove "-lmpfr"
from the build system in a few places. Since configure.ac already
calls AC_CHECK_LIB(mpfr,...) unconditionally, the "-lmpfr" flag will
always be present in LIBS.