{fplll,tests}/Makefile.am: drop LIBS from LIBADD and LDADD.
The contents of the $LIBS variable is already appended to the linker
command, so there's no need to add it ourselves in fplll/Makefile.am
and tests/Makefile.am. This fixes the double-appearance of some flags
in (for example)

  /bin/bash ../libtool  --tag=CXX   --mode=link g++ -std=c++11 -pthread
  -O0 -no-undefined -version-info 6:2:0 -pthread -Wl,-O1 -Wl,--as-needed
  -o libfplll.la ... -lqd -lmpfr -lgmp   -lpthread -lqd -lmpfr -lgmp

where "-lqd -lmpfr -lgmp" is repeated. After this commit, those flags
only appear once.